chore(types): fix types in more tests
