+++ packages/graphile-build/__tests__/Live.test.ts
@@ -3,7 +3,10 @@ import {
   LiveCoordinator,
   LiveProvider,
   makeAsyncIteratorFromMonitor,
-} from "../dist/Live";
+  LiveSubscriptionCallback,
+  LivePredicate,
+  LiveSubscriptionReleaser,
+} from "../src/Live";
 
 // TODO: this file needs a *lot* more tests!
 
@@ -30,7 +33,11 @@ class DummySource extends LiveSource {
     this.recordListeners = [];
   }
 
-  subscribeCollection(callback, collectionIdentifier, predicate) {
+  subscribeCollection(
+    callback: LiveSubscriptionCallback,
+    collectionIdentifier: any,
+    predicate?: LivePredicate,
+  ): LiveSubscriptionReleaser | null {
     const entry = [callback, collectionIdentifier, predicate];
     this.collectionListeners.push(entry);
     return () => {
@@ -41,7 +48,11 @@ class DummySource extends LiveSource {
     };
   }
 
-  subscribeRecord(callback, collectionIdentifier, recordIdentifier) {
+  subscribeRecord(
+    callback: LiveSubscriptionCallback,
+    collectionIdentifier: any,
+    recordIdentifier: any,
+  ): LiveSubscriptionReleaser | null {
     const entry = [callback, collectionIdentifier, recordIdentifier];
     this.recordListeners.push(entry);
     return () => {
@@ -52,7 +63,7 @@ class DummySource extends LiveSource {
     };
   }
 
-  triggerRecord(collectionIdentifier, recordIdentifier) {
+  triggerRecord(collectionIdentifier: symbol, recordIdentifier: symbol) {
     for (const rl of this.recordListeners) {
       const [rlCallback, rlCollectionIdentifier, rlRecordIdentifier] = rl;
       if (
@@ -117,7 +128,7 @@ Array [
   // another update it won't do anything
   dummySource.triggerRecord(collection, record);
   // So if we return
-  asyncIterator.return();
+  asyncIterator.return?.();
   // Then done should be true immediately (no intermediate values generated by the above triggerRecord)
   {
     const { value, done } = await asyncIterator.next();
