+++ packages/graphile-build/__tests__/fieldData.test.ts
@@ -1,3 +1,4 @@
+import "../global";
 import {
   graphql,
   GraphQLObjectType,
